refactor(perp): move keeper files to v1 directory
