+++ x/perp/keeper/v1/amm_test.go
@@ -13,7 +13,7 @@ import (
 	"github.com/NibiruChain/nibiru/x/common/testutil"
 	"github.com/NibiruChain/nibiru/x/common/testutil/testapp"
 	perpammtypes "github.com/NibiruChain/nibiru/x/perp/amm/types"
-	"github.com/NibiruChain/nibiru/x/perp/keeper"
+	"github.com/NibiruChain/nibiru/x/perp/keeper/v1"
 	types "github.com/NibiruChain/nibiru/x/perp/types/v1"
 )
