+++ x/perp/simulation/operations.go
@@ -17,7 +17,7 @@ import (
 	"github.com/NibiruChain/nibiru/x/common/denoms"
 	perpammtypes "github.com/NibiruChain/nibiru/x/perp/amm/types"
 	pooltypes "github.com/NibiruChain/nibiru/x/perp/amm/types"
-	"github.com/NibiruChain/nibiru/x/perp/keeper"
+	"github.com/NibiruChain/nibiru/x/perp/keeper/v1"
 	types "github.com/NibiruChain/nibiru/x/perp/types/v1"
 )
